Refactor imports: move PropertyPrinter to utils package in Reservation, Schedule, and Table models
